TOT: moves file reading into resources.cpp
